Move rsc.classpath to rsc.symtab
